Editing text reflecting in the list inside ViewModel
There's a bug of keyboard because each keystroke redraws the whole list. To avoid this I can make each change in the text only change the StoryStep locally and the change is propagated to the view model, but it doesn't propagate it to the UI right away and buffers it to be changed latter.